divide test (#721)
